html/browsers/browsing-the-web/back-forward-cache/events.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
for (const originType of ['SameOrigin', 'SameSite', 'CrossSite']) {
  promise_test(async t => {
    const pageA = new ExecutorTarget();

    window.open(
      executorPath + pageA.uuid + '&events=pagehide,pageshow,load',
      '_blank', 'noopener');

    // Navigate to a page that immediately triggers back navigation.
    const backUrl = eval(`origin${originType}`) + backPath;
    await pageA.exec(`
        prepareNavigation();
        location.href = '${backUrl}';
    `);

    await assert_bfcached(pageA);

    assert_array_equals(
      await pageA.eval(`getRecordedEvents()`),
      [
        'window.load',
        'window.pageshow',
        'window.pagehide.persisted',
        'window.pageshow.persisted'
      ]);
  }, `Events fired (window.open + noopener, ${originType})`);
}
</script>
